♻  refactor: Split utils into docs / github / all
