+++ src/siesta/utils/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2025 Entalpic
+"""A set of utilities to help the ``siesta`` CLI."""
